Restructure to support more complex types
