+++ lib/App.ts
@@ -2,8 +2,9 @@ import { ApiPromiseInterface } from "@polkadot/api/promise/types"
 import { getTypeRegistry } from "@polkadot/types"
 import { TypeRegistry } from "@polkadot/types/codec/typeRegistry"
 import { GraphQLServer } from "./GraphQLServer"
-
 import { GraphQLServerMetadataConfig } from "./GraphQLServerMetadataConfig"
+import { TypeClassifier } from "./TypeClassifier"
+import { QueryResolver } from "./QueryResolver"
 
 const log = require("npmlog")
 
@@ -19,8 +20,8 @@ export class App {
 
     public async start() {
         const config = new GraphQLServerMetadataConfig(
-            this.api,
-            this.typeRegistry,
+            new QueryResolver(this.api),
+            new TypeClassifier(this.typeRegistry),
             this.api.runtimeMetadata.asV3,
         )
         const server = new GraphQLServer(config)
